pyproject.toml
[project]
name = "work-hours-bot"
version = "1.0.0"
description = "Bot used to track working hours."
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "aiogram>=3.26.0",
    "aiosqlite>=0.22.1",
    "sqlalchemy>=2.0.48",
    "sqlalchemy-utc>=0.14.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "python-dotenv>=1.2.2",
]
